+++ changelog.md
@@ -1,9 +1,14 @@
 # Change Log
 
+## v1.1.0 - 040220
+### New functions
+1. friftbatterytable - View details of furoma rift battery levels, repair, cost, etc.
+2. vriftaugment - View details of valor rift augmentations.
+
 ## v1.0.4 - 040220
-1. Improved efficiencry for vriftfragcore function
+1. Improved efficiency for vriftfragcore function
 ### Bug fix
-1. Fix bug in vriftfragcore function outputing wrong floor and values.<br/>
+1. Fix bug in vriftfragcore function outputing wrong floor and values.
 
 ## v1.0.3 - 030220
 1. Removal of test messages on backend